数据库修改为sqlite

+++ model/req/common.go
@@ -1,7 +1,7 @@
 package req
 
-// 分页
+// PageInfo 分页
 type PageInfo struct {
-	PageIndex int64 `form:"pageIndex" binding:"required"`
-	PageSize  int64 `form:"pageSize" binding:"required"`
+	PageIndex int `form:"pageIndex" binding:"required"`
+	PageSize  int `form:"pageSize" binding:"required"`
 }
